+++ app/views/companion/blockchain/SendAsset.scala
@@ -15,6 +15,6 @@ object SendAsset {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(from: String, to: String, pegHash: String, gas: Int, mode: String, password: String)
+  case class Data(from: String, to: String, pegHash: String, gas: Long, mode: String, password: String)
 
 }
