+++ app/views/companion/master/AcceptNegotiationRequest.scala
@@ -12,6 +12,6 @@ object AcceptNegotiationRequest {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(id: String, gas: Int, password: String)
+  case class Data(id: String, gas: Long, password: String)
 
 }
