+++ app/views/companion/master/SellerExecuteOrder.scala
@@ -12,6 +12,6 @@ object SellerExecuteOrder {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(orderID: String, gas: Int, password: String)
+  case class Data(orderID: String, gas: Long, password: String)
 
 }
